requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Werkzeug==2.3.6
requests==2.31.0
gunicorn==21.2.0
firebase-admin==6.2.0
psycopg2-binary==2.9.7 